Update: MTU manage, cleanup
